refactor: move packages to io.github.joshuakfarrar.apollo
Aligns the Scala package namespace with the Maven groupId before
first publication, while renaming is still free.
