+++ apollo-core/src/io/github/joshuakfarrar/apollo/core/ConfirmationService.scala
@@ -1,0 +1,9 @@
+package io.github.joshuakfarrar.apollo.core
+
+import cats.data.{EitherT, OptionT}
+
+trait ConfirmationService[F[_], U, I] {
+  def createConfirmation(user: U): EitherT[F, Throwable, String]
+
+  def confirmByCode(code: String): OptionT[F, Throwable]
+}
