+++ apollo-core/src/io/github/joshuakfarrar/apollo/core/HasEmail.scala
@@ -1,0 +1,4 @@
+package io.github.joshuakfarrar.apollo.core
+
+trait HasEmail[U]:
+  def email(user: U): String
