+++ apollo-core/src/io/github/joshuakfarrar/apollo/core/HasId.scala
@@ -1,0 +1,4 @@
+package io.github.joshuakfarrar.apollo.core
+
+trait HasId[U, I]:
+  def id(user: U): I
