+++ apollo-core/src/io/github/joshuakfarrar/apollo/core/HasPassword.scala
@@ -1,0 +1,4 @@
+package io.github.joshuakfarrar.apollo.core
+
+trait HasPassword[U]:
+  def password(user: U): String
