+++ apollo-core/src/io/github/joshuakfarrar/apollo/core/util.scala
@@ -1,0 +1,11 @@
+package io.github.joshuakfarrar.apollo.core
+
+import cats.Monad
+import cats.effect.std.Random
+import cats.implicits.*
+
+object util:
+  def generateAlphaNumericString[F[_]: Monad](length: Int)(using
+      R: Random[F]
+  ): F[String] =
+    R.nextAlphaNumeric.replicateA(length).map(_.mkString)
