--- apollo-core/src/me/joshuakfarrar/apollo/core/HasEmail.scala
@@ -1,4 +1,0 @@
-package me.joshuakfarrar.apollo.core
-
-trait HasEmail[U]:
-  def email(user: U): String
