--- apollo-core/src/me/joshuakfarrar/apollo/core/HasId.scala
@@ -1,4 +1,0 @@
-package me.joshuakfarrar.apollo.core
-
-trait HasId[U, I]:
-  def id(user: U): I
