--- apollo-core/src/me/joshuakfarrar/apollo/core/HasPassword.scala
@@ -1,4 +1,0 @@
-package me.joshuakfarrar.apollo.core
-
-trait HasPassword[U]:
-  def password(user: U): String
