--- apollo-http4s/src/me/joshuakfarrar/apollo/http4s/FlashMiddleware.scala
@@ -1,31 +1,0 @@
-package me.joshuakfarrar.apollo.http4s
-
-import cats.data.Kleisli
-import cats.effect.Sync
-import org.http4s.headers.Cookie
-import org.http4s.{HttpDate, HttpRoutes, Request, ResponseCookie}
-
-import java.time.Instant
-
-object FlashMiddleware {
-  def httpRoutes[F[_]: Sync](service: HttpRoutes[F]): HttpRoutes[F] = Kleisli {
-    (req: Request[F]) =>
-      service(req).map { response =>
-        val hasFlashCookie = req.headers
-          .get[Cookie]
-          .exists(_.values.exists(_.name == "flash"))
-
-        if (hasFlashCookie) {
-          val deleteFlashCookie = ResponseCookie(
-            name = "flash",
-            content = "",
-            expires = Some(HttpDate.unsafeFromInstant(Instant.EPOCH)),
-            path = Some("/")
-          )
-          response.addCookie(deleteFlashCookie)
-        } else {
-          response
-        }
-      }
-  }
-}
